requirements.txt
vllm==0.9.0
math-verify
ms-swift==3.6.0
deepspeed==0.16.9
liger-kernel
flash-attn
ipykernel
gpustat
fastapi
qwen-vl-utils
FlagEmbedding
retrying
scikit-learn
openai
tqdm_joblib
tenacity
matplotlib
pytrec-eval-terrier
